Process only files related to master file during upload translation (#4)
When initializing a translation file, the process looped over all files in the directory of the master file. If there were others files than the ones related to the given master file, they were processed as well.
This fix allows to only process translation files related to the given master file.